+++ apps/ensadmin/src/lib/system-clock.ts
@@ -1,8 +1,0 @@
-import { HighResolutionSyncedClock, SyncedClock } from "@/lib/synced-clock";
-
-/**
- * Synced System Clock
- *
- * There is just one instance of it in ENSAdmin.
- */
-export const systemClock: SyncedClock = new HighResolutionSyncedClock();