+++ apps/ensindexer/src/plugins/registrars/basenames/handlers/Basenames_RegistrarController.ts
@@ -63,7 +63,6 @@ export default function () {
 
       await handleRegistrarControllerEvent(context, {
         id,
-        subregistryId,
         node,
         pricing,
         referral,
@@ -94,7 +93,6 @@ export default function () {
 
       await handleRegistrarControllerEvent(context, {
         id,
-        subregistryId,
         node,
         pricing,
         referral,
@@ -121,7 +119,6 @@ export default function () {
 
       await handleRegistrarControllerEvent(context, {
         id,
-        subregistryId,
         node,
         pricing,
         referral,
@@ -152,7 +149,6 @@ export default function () {
 
       await handleRegistrarControllerEvent(context, {
         id,
-        subregistryId,
         node,
         pricing,
         referral,
@@ -179,7 +175,6 @@ export default function () {
 
       await handleRegistrarControllerEvent(context, {
         id,
-        subregistryId,
         node,
         pricing,
         referral,