+++ packages/namehash-ui/src/utils/blockExplorers.ts
@@ -2,7 +2,7 @@ import type { Address, Hash } from "viem";
 
 import type { ChainId } from "@ensnode/ensnode-sdk";
 
-import { SUPPORTED_CHAINS } from "@/utils/chains.ts";
+import { SUPPORTED_CHAINS } from "./chains";
 
 /**
  * Gets the "base" block explorer URL for a given {@link ChainId}
@@ -50,4 +50,17 @@ export const getBlockExplorerTransactionDetailsUrl = (
   if (!chainBlockExplorer) return null;
 
   return new URL(`tx/${transactionHash}`, chainBlockExplorer.toString());
+};
+
+/**
+ * Gets the block explorer URL for a specific block on a specific chainId
+ *
+ * @returns complete block explorer URL for a specific block on a specific chainId,
+ * or null if the referenced chain doesn't have a known block explorer
+ */
+export const getBlockExplorerBlockUrl = (chainId: ChainId, blockNumber: number): URL | null => {
+  const chainBlockExplorer = getBlockExplorerUrl(chainId);
+  if (!chainBlockExplorer) return null;
+
+  return new URL(`block/${blockNumber}`, chainBlockExplorer.toString());
 };