+++ apps/ensapi/CHANGELOG.md
@@ -1,5 +1,37 @@
 # ensapi
 
+## 1.5.1
+
+### Patch Changes
+
+- [#1529](https://github.com/namehash/ensnode/pull/1529) [`b25605b`](https://github.com/namehash/ensnode/commit/b25605b49c09d57dae44ff53092303eb7330df85) Thanks [@notrab](https://github.com/notrab)! - Add production and testnet servers to OpenAPI spec
+
+  The OpenAPI specification now includes the following servers:
+
+  - https://api.alpha.ensnode.io (ENSNode Alpha - Mainnet)
+  - https://api.alpha-sepolia.ensnode.io (ENSNode Alpha - Sepolia Testnet)
+  - localhost (Local Development)
+
+- [#1537](https://github.com/namehash/ensnode/pull/1537) [`63617fa`](https://github.com/namehash/ensnode/commit/63617fa827daa4bd7761f482812daf7b507da3d2) Thanks [@tk-o](https://github.com/tk-o)! - Updates Node.js runtime to the current LTS version (v24).
+
+- Updated dependencies []:
+  - @ensnode/datasources@1.5.1
+  - @ensnode/ensnode-schema@1.5.1
+  - @ensnode/ponder-subgraph@1.5.1
+  - @ensnode/ensnode-sdk@1.5.1
+  - @namehash/ens-referrals@1.5.1
+
+## 1.5.0
+
+### Patch Changes
+
+- Updated dependencies [[`dc7e07f`](https://github.com/namehash/ensnode/commit/dc7e07f6e69e30d597a871b79bd2c6876de9f8cc)]:
+  - @ensnode/ensnode-schema@1.5.0
+  - @ensnode/datasources@1.5.0
+  - @ensnode/ponder-subgraph@1.5.0
+  - @ensnode/ensnode-sdk@1.5.0
+  - @namehash/ens-referrals@1.5.0
+
 ## 1.4.0
 
 ### Minor Changes