+++ apps/fallback-ensapi/CHANGELOG.md
@@ -1,5 +1,21 @@
 # fallback-ensapi
 
+## 1.5.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ensnode/datasources@1.5.1
+  - @ensnode/ensnode-sdk@1.5.1
+
+## 1.5.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ensnode/datasources@1.5.0
+  - @ensnode/ensnode-sdk@1.5.0
+
 ## 1.4.0
 
 ### Minor Changes