+++ packages/datasources/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @ensnode/ens-deployments
 
+## 1.5.1
+
+## 1.5.0
+
 ## 1.4.0
 
 ### Minor Changes