+++ packages/ens-referrals/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @namehash/ens-referrals
 
+## 1.5.1
+
+## 1.5.0
+
 ## 1.4.0
 
 ### Minor Changes