+++ packages/ensrainbow-sdk/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @ensnode/ensrainbow-sdk
 
+## 1.5.1
+
+## 1.5.0
+
 ## 1.4.0
 
 ### Patch Changes